feat: complete modularization and rename mvvm to repl
- Organized commands into logical submodules (movement, editing, mode, pane, request)
- Organized models into focused submodules (buffer, editor, request, response)
- Organized events into logical submodules (types, model events, view events, event bus)
- Renamed mvvm module to repl to better reflect its purpose as a REPL interface
- Maintained same public API with zero breaking changes
- Added comprehensive tests for all new modules (69 total tests passing)
- All code is clippy-clean and properly formatted
- Architecture is now highly scalable and maintainable
