+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "blueline"
-version = "0.9.0"
+version = "0.10.0"
 authors = ["Sam Wisely <samwisely75@gmail.com>"]
 description = "A lightweight, profile-based HTTP client with REPL interface"
 edition = "2021"
